Fix type of pagesize, size in mmapForLinker
Caught by format string checking; "bytes" used to be used (before
69db8adf9c) where "size" is now, and changing "pagesize" and "size" to
size_t looks more correct than changing the format string.

+++ rts/Linker.c
@@ -1974,7 +1974,7 @@ mmapForLinker (size_t bytes, nat flags, int fd)
 {
    void *map_addr = NULL;
    void *result;
-   int pagesize, size;
+   size_t pagesize, size;
    static nat fixed = 0;
 
    IF_DEBUG(linker, debugBelch("mmapForLinker: start\n"));
